adding conditioning noise
